+++ changelog.md
@@ -4,6 +4,7 @@
 - added `initOptions` method to `Grid`
 - added `invalidateVisuals` method to `Grid`
 - changed way how is grid initialized, more consistent behavior, ctor => initOptions => createPlugins => setPluginOptions => initPluginOptions => initializeGrid => initializePlugins, same works for nested plugins
+- added extension method for reinitialization of grid options `reinitializeOptions`
 
 ## Version 5.0.0-beta.11
 - updated `PagingAbstractComponent` and `BasicPagingComponent`, all `private` members changed to `protected`